+++ docs/changelog.md
@@ -4,6 +4,12 @@ ProTaskinate changelog
 
 This format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- User can now view updated_at for all tasks
+
 ## [v0.2.0-beta] - 2024-10-01
 
 ### Added